// src/app/models/job-list-model.ts
export class JobList {
    job_id: number;
    master_job_id: number;
    facility_address: string;
    onsite_phone: boolean;
    onsite_person: String;
    property_unit: string;
    client_id: number;
    FacilityAddress: string;
    FacilityCity: string;
    FacilityState: string;
    FacilityZip: string;
    technician: string;
    date_scheduled: string;
    StartDateTime: string;
    EndDateTime: string;
    service_id: string;
}